VideoThumbnailWidget implementation
